+++ packages/exchange/src/utils/valueTransformers.ts
@@ -2,6 +2,6 @@ import { ValueTransformer } from 'typeorm';
 import BN from 'bn.js';
 
 export const BNTransformer: ValueTransformer = {
-    to: (entityValue: BN) => entityValue.toString(10),
+    to: (entityValue: BN) => entityValue?.toString(10),
     from: (databaseValue: string): BN => new BN(databaseValue)
 };